+++ src/main/java/uk/gov/service/notify/domain/NotificationType.java
@@ -1,7 +1,10 @@
 package uk.gov.service.notify.domain;
 
+import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
+
 public enum NotificationType {
     sms,
     email,
-    letter
+    letter,
+    @JsonEnumDefaultValue UNKNOWN
 }