+++ src/main/java/uk/gov/service/notify/domain/Postage.java
@@ -1,12 +1,14 @@
 package uk.gov.service.notify.domain;
 
+import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public enum Postage {
     @JsonProperty("first") FIRST("first"),
     @JsonProperty("second") SECOND("second"),
     @JsonProperty("europe") EUROPE("europe"),
-    @JsonProperty("rest-of-world") REST_OF_WORLD("rest-of-world");
+    @JsonProperty("rest-of-world") REST_OF_WORLD("rest-of-world"),
+    @JsonEnumDefaultValue UNKNOWN("unknown");
 
     private final String postage;