+++ src/test/java/uk/gov/service/notify/ValidatingJsonMapperTest.java
@@ -1,5 +1,6 @@
 package uk.gov.service.notify;
 
+import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import jakarta.validation.constraints.NotNull;
@@ -16,18 +17,32 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class ValidatingJsonMapperTest {
 
+    private enum TestEnum {
+        KNOWN,
+        @JsonEnumDefaultValue UNKNOWN
+    }
+
     private static class TestClassForSerialization {
         @NotNull
         @PastOrPresent
         private final ZonedDateTime time;
+        @NotNull
+        private final TestEnum testEnum;
 
-        public TestClassForSerialization(@JsonProperty("time") ZonedDateTime time) {
+        public TestClassForSerialization(@JsonProperty("time") ZonedDateTime time,
+                                         @JsonProperty("enum") TestEnum testEnum) {
             this.time = time;
+            this.testEnum = testEnum;
         }
 
         @JsonProperty("time")
         public ZonedDateTime getTime() {
             return time;
+        }
+
+        @JsonProperty("enum")
+        public TestEnum getTestEnum() {
+            return testEnum;
         }
     }
 
@@ -38,11 +53,11 @@ class ValidatingJsonMapperTest {
         ValidatingJsonMapper mapper = new ValidatingJsonMapper(options);
         String expected = "2024-05-10T16:40:14Z";
         ZonedDateTime time = ZonedDateTime.parse(expected);
-        TestClassForSerialization test = new TestClassForSerialization(time);
+        TestClassForSerialization test = new TestClassForSerialization(time, TestEnum.KNOWN);
 
         String string = mapper.writeValueAsString(test);
 
-        assertThat(string).isEqualTo("{\"time\":\""+expected+"\"}");
+        assertThat(string).isEqualTo("{\"time\":\""+expected+"\",\"enum\":\"KNOWN\"}");
     }
 
     @Test
@@ -52,7 +67,7 @@ class ValidatingJsonMapperTest {
         ValidatingJsonMapper mapper = new ValidatingJsonMapper(options);
         String expected = "3024-05-10T16:40:14Z";
         ZonedDateTime time = ZonedDateTime.parse(expected);
-        TestClassForSerialization test = new TestClassForSerialization(time);
+        TestClassForSerialization test = new TestClassForSerialization(time, TestEnum.KNOWN);
 
         NotificationClientException exception = assertThrows(NotificationClientException.class,
                 () -> mapper.writeValueAsString(test));
@@ -67,11 +82,11 @@ class ValidatingJsonMapperTest {
         ValidatingJsonMapper mapper = new ValidatingJsonMapper(options);
         String expected = "3024-05-10T16:40:14Z";
         ZonedDateTime time = ZonedDateTime.parse(expected);
-        TestClassForSerialization test = new TestClassForSerialization(time);
+        TestClassForSerialization test = new TestClassForSerialization(time, TestEnum.KNOWN);
 
         String string = mapper.writeValueAsString(test);
 
-        assertThat(string).isEqualTo("{\"time\":\""+expected+"\"}");
+        assertThat(string).isEqualTo("{\"time\":\""+expected+"\",\"enum\":\"KNOWN\"}");
     }
 
     @Test
@@ -80,11 +95,12 @@ class ValidatingJsonMapperTest {
                 .setOption(NotificationClientOptions.Options.VALIDATION_SKIP, "false");
         ValidatingJsonMapper mapper = new ValidatingJsonMapper(options);
         String expectedTime = "2024-05-10T16:40:14Z";
-        String expected = "{\"time\":\""+expectedTime+"\"}";
+        String expected = "{\"time\":\""+expectedTime+"\",\"enum\":\"KNOWN\"}";
 
         TestClassForSerialization actual = mapper.readValue(new ByteArrayInputStream(expected.getBytes(StandardCharsets.UTF_8)), TestClassForSerialization.class);
 
         assertThat(actual.getTime()).isEqualTo(expectedTime);
+        assertThat(actual.getTestEnum()).isEqualTo(TestEnum.KNOWN);
     }
 
     @Test
@@ -93,7 +109,7 @@ class ValidatingJsonMapperTest {
                 .setOption(NotificationClientOptions.Options.VALIDATION_SKIP, "false");
         ValidatingJsonMapper mapper = new ValidatingJsonMapper(options);
         String expectedTime = "3024-05-10T16:40:14Z";
-        String expected = "{\"time\":\""+expectedTime+"\"}";
+        String expected = "{\"time\":\""+expectedTime+"\",\"enum\":\"foo\"}";
 
         NotificationClientException exception = assertThrows(NotificationClientException.class,
                 () -> mapper.readValue(new ByteArrayInputStream(expected.getBytes(StandardCharsets.UTF_8)), TestClassForSerialization.class));
@@ -107,10 +123,11 @@ class ValidatingJsonMapperTest {
                 .setOption(NotificationClientOptions.Options.VALIDATION_SKIP, "true");
         ValidatingJsonMapper mapper = new ValidatingJsonMapper(options);
         String expectedTime = "3024-05-10T16:40:14Z";
-        String expected = "{\"time\":\""+expectedTime+"\"}";
+        String expected = "{\"time\":\""+expectedTime+"\",\"enum\":\"foo\"}";
 
         TestClassForSerialization actual = mapper.readValue(new ByteArrayInputStream(expected.getBytes(StandardCharsets.UTF_8)), TestClassForSerialization.class);
 
         assertThat(actual.getTime()).isEqualTo(expectedTime);
+        assertThat(actual.getTestEnum()).isEqualTo(TestEnum.UNKNOWN);
     }
 }